[Statie] move Amp to standalone subpackage
