+++ composer.json
@@ -53,6 +53,7 @@
             "Symplify\\EasyCodingStandard\\FixerRunner\\": "packages/EasyCodingStandard/packages/FixerRunner/src",
             "Symplify\\PackageBuilder\\": "packages/PackageBuilder/src",
             "Symplify\\Statie\\": "packages/Statie/src",
+            "Symplify\\Statie\\Amp\\": "packages/Statie/packages/Amp/src",
             "Symplify\\Statie\\FlatWhite\\": "packages/Statie/packages/FlatWhite/src",
             "Symplify\\Statie\\SimilarPosts\\": "packages/Statie/packages/SimilarPosts/src",
             "Symplify\\SymbioticController\\": "packages/SymbioticController/src"
@@ -68,6 +69,7 @@
             "Symplify\\EasyCodingStandard\\FixerRunner\\Tests\\": "packages/EasyCodingStandard/packages/FixerRunner/tests",
             "Symplify\\PackageBuilder\\Tests\\": "packages/PackageBuilder/tests",
             "Symplify\\Statie\\Tests\\": "packages/Statie/tests",
+            "Symplify\\Statie\\Amp\\Tests\\": "packages/Statie/packages/Amp/tests",
             "Symplify\\Statie\\FlatWhite\\Tests\\": "packages/Statie/packages/FlatWhite/tests",
             "Symplify\\Statie\\SimilarPosts\\Tests\\": "packages/Statie/packages/SimilarPosts/tests",
             "Symplify\\SymbioticController\\Tests\\": "packages/SymbioticController/tests",
