+++ packages/Statie/composer.json
@@ -35,6 +35,7 @@
     "autoload": {
         "psr-4": {
             "Symplify\\Statie\\": "src",
+            "Symplify\\Statie\\Amp\\": "packages/Amp/src",
             "Symplify\\Statie\\FlatWhite\\": "packages/FlatWhite/src",
             "Symplify\\Statie\\SimilarPosts\\": "packages/SimilarPosts/src"
         }
@@ -42,6 +43,7 @@
     "autoload-dev": {
         "psr-4": {
             "Symplify\\Statie\\Tests\\": "tests",
+            "Symplify\\Statie\\Amp\\Tests\\": "packages/Amp/tests",
             "Symplify\\Statie\\FlatWhite\\Tests\\": "packages/FlatWhite/tests",
             "Symplify\\Statie\\SimilarPosts\\Tests\\": "packages/SimilarPosts/tests"
         }
