+++ packages/Statie/packages/Amp/src/AmpLinkFileDecorator.php
