+++ packages/Statie/src/config/services.yml
@@ -6,6 +6,8 @@ services:
     Symplify\Statie\:
         resource: ../../src
         exclude: ../../src/{Exception,Renderable/File/File.php,Renderable/File/PostFile.php}
+    Symplify\Statie\Amp\:
+        resource: ../../packages/Amp/src
     Symplify\Statie\FlatWhite\:
         resource: ../../packages/FlatWhite/src
     Symplify\Statie\SimilarPosts\:
